Renamed and moved DollaDataStubs.
