+++ src/test/java/dolla/DollaDataStubs/DollaDataStub1.java
@@ -1,4 +1,4 @@
-package dolla.command.view;
+package dolla.DollaDataStubs;
 
 import dolla.model.BillList;
 import dolla.model.DebtList;
@@ -12,15 +12,22 @@ import dolla.model.ShortcutList;
 import java.time.LocalDate;
 import java.util.ArrayList;
 
-public class ViewDollaDataStub3 extends DollaData {
+public class DollaDataStub1 extends DollaData {
+
+    private String mode = MODE_DOLLA;
+    private EntryList entryList;
+    private DebtList debtList;
+    private LimitList limitList;
+    private ShortcutList shortcutList;
+    private BillList billList;
 
     private Entry createNewEntry1() {
-        return new Entry("expense", 200, "Expense Description",
-                LocalDate.parse("1111-11-11"));
+        return new Entry("expense", 100, "Expense Description",
+                LocalDate.parse("2001-12-03"));
     }
 
     private Entry createNewEntry2() {
-        return new Entry("income", 100, "Income Description",
+        return new Entry("income", 200, "Income Description",
                 LocalDate.parse("1111-11-11"));
     }
 
@@ -29,18 +36,17 @@ public class ViewDollaDataStub3 extends DollaData {
                 LocalDate.parse("1000-01-01"));
     }
 
+    /**
+     * Creates an instance of DollaData to store and manipulate data.
+     */
     private ArrayList<Record> createNewEntryArrayList() {
         ArrayList<Record> newRecordList = new ArrayList<Record>();
         newRecordList.add(createNewEntry1());
         newRecordList.add(createNewEntry2());
-        newRecordList.add(createNewEntry3());
         return newRecordList;
     }
 
-    /**
-     * Creates an instance of DollaData to store and manipulate data.
-     */
-    public ViewDollaDataStub3() {
+    public DollaDataStub1() {
         this.entryList = new EntryList(createNewEntryArrayList());
         this.limitList = new LimitList(new ArrayList<Record>());
         this.debtList = new DebtList(new ArrayList<Record>());
