+++ src/test/java/dolla/command/view/ViewCommandTest.java
@@ -2,6 +2,10 @@ package dolla.command.view;
 
 import java.time.LocalDate;
 
+import dolla.DollaDataStubs.DollaDataStub0;
+import dolla.DollaDataStubs.DollaDataStub1;
+import dolla.DollaDataStubs.DollaDataStub2;
+import dolla.DollaDataStubs.DollaDataStub3;
 import dolla.model.DollaData;
 import org.junit.jupiter.api.Test;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -10,7 +14,7 @@ public class ViewCommandTest {
 
     @Test
     public void execute_emptyList_zero() {
-        DollaData dollaData = new ViewDollaDataStub0();
+        DollaData dollaData = new DollaDataStub0();
         ViewTodayCommand viewCommand = new ViewTodayCommand();
         viewCommand.execute(dollaData);
         double expectedSum = 0;
@@ -19,7 +23,7 @@ public class ViewCommandTest {
 
     @Test
     public void execute_noRelevantDates_zero() {
-        DollaData dollaData = new ViewDollaDataStub1();
+        DollaData dollaData = new DollaDataStub1();
         LocalDate inputDate = LocalDate.parse("1001-12-03");
         ViewDateCommand viewCommand = new ViewDateCommand(inputDate);
         viewCommand.execute(dollaData);
@@ -29,7 +33,7 @@ public class ViewCommandTest {
 
     @Test
     public void execute_positiveOverall_positive() {
-        DollaData dollaData = new ViewDollaDataStub2();
+        DollaData dollaData = new DollaDataStub2();
         LocalDate inputDate = LocalDate.parse("1111-11-11");
         ViewDateCommand viewCommand = new ViewDateCommand(inputDate);
         viewCommand.execute(dollaData);
@@ -39,7 +43,7 @@ public class ViewCommandTest {
 
     @Test
     public void execute_positiveOverall_negative() {
-        DollaData dollaData = new ViewDollaDataStub3();
+        DollaData dollaData = new DollaDataStub3();
         LocalDate inputDate = LocalDate.parse("1111-11-11");
         ViewDateCommand viewCommand = new ViewDateCommand(inputDate);
         viewCommand.execute(dollaData);
